+++ engine_py/src/api.rs
@@ -1,650 +1,1 @@
-use engine_core::ecs::world::World;
-use engine_core::map::{Map, SquareGridMap};
-use engine_core::systems::job::JobSystem;
-use engine_core::systems::standard::{DamageAll, MoveAll, MoveDelta, ProcessDeaths, ProcessDecay};
-use pyo3::exceptions::{PyIOError, PyValueError};
-use pyo3::prelude::*;
-use pyo3::types::{PyAny, PyDict};
-use serde_json::Value;
-use serde_pyobject::{from_pyobject, to_pyobject};
-use std::cell::RefCell;
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::sync::{Arc, Mutex};
-
-use crate::event_bus;
-use crate::system_bridge::SystemBridge;
-
-#[pyclass(unsendable)]
-pub struct PyWorld {
-    pub inner: Rc<RefCell<World>>,
-    pub systems: Rc<SystemBridge>,
-}
-
-#[pymethods]
-impl PyWorld {
-    #[new]
-    #[pyo3(signature = (schema_dir=None))]
-    fn new(schema_dir: Option<String>) -> PyResult<Self> {
-        use engine_core::ecs::registry::ComponentRegistry;
-        use engine_core::ecs::schema::load_schemas_from_dir;
-        use std::path::PathBuf;
-
-        let schema_path = match schema_dir {
-            Some(dir) => PathBuf::from(dir),
-            None => PathBuf::from("engine/assets/schemas"),
-        };
-
-        let schemas = load_schemas_from_dir(&schema_path).map_err(|e| {
-            PyValueError::new_err(format!(
-                "Failed to load schemas from {:?}: {e}",
-                schema_path
-            ))
-        })?;
-
-        let mut registry = ComponentRegistry::new();
-        for (_name, schema) in schemas {
-            registry.register_external_schema(schema);
-        }
-
-        let mut world = World::new(Arc::new(Mutex::new(registry)));
-
-        // Always initialize a map for the world (so add_cell and movement will work)
-        let grid = SquareGridMap::new();
-        let map = Map::new(Box::new(grid));
-        world.map = Some(map);
-
-        world.register_system(MoveAll {
-            delta: MoveDelta::Square {
-                dx: 1,
-                dy: 0,
-                dz: 0,
-            },
-        });
-        world.register_system(DamageAll { amount: 1.0 });
-        world.register_system(ProcessDeaths);
-        world.register_system(ProcessDecay);
-        world.register_system(JobSystem::default());
-        Ok(PyWorld {
-            inner: Rc::new(RefCell::new(world)),
-            systems: Rc::new(SystemBridge {
-                systems: RefCell::new(HashMap::new()),
-            }),
-        })
-    }
-
-    // --- ECS methods ---
-
-    fn spawn_entity(&self) -> u32 {
-        let mut world = self.inner.borrow_mut();
-        world.spawn_entity()
-    }
-
-    fn despawn_entity(&self, entity_id: u32) {
-        let mut world = self.inner.borrow_mut();
-        world.despawn_entity(entity_id);
-        world.entities.retain(|&e| e != entity_id);
-    }
-
-    fn set_component(&self, entity_id: u32, name: String, value: Bound<'_, PyAny>) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let json_value: Value = from_pyobject(value)?;
-        world
-            .set_component(entity_id, &name, json_value)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    fn get_component(
-        &self,
-        py: Python<'_>,
-        entity_id: u32,
-        name: String,
-    ) -> PyResult<Option<PyObject>> {
-        let world = self.inner.borrow_mut();
-        if let Some(val) = world.get_component(entity_id, &name) {
-            let py_obj = to_pyobject(py, val)?;
-            Ok(Some(py_obj.into()))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn remove_component(&self, entity_id: u32, name: String) {
-        let mut world = self.inner.borrow_mut();
-        if let Some(comps) = world.components.get_mut(&name) {
-            comps.remove(&entity_id);
-        }
-    }
-
-    fn get_entities_with_component(&self, name: String) -> PyResult<Vec<u32>> {
-        let world = self.inner.borrow_mut();
-        Ok(world.get_entities_with_component(&name))
-    }
-
-    fn get_entities_with_components(&self, names: Vec<String>) -> Vec<u32> {
-        let world = self.inner.borrow_mut();
-        let name_refs: Vec<&str> = names.iter().map(|s| s.as_str()).collect();
-        world.get_entities_with_components(&name_refs)
-    }
-
-    fn get_entities(&self) -> PyResult<Vec<u32>> {
-        let world = self.inner.borrow_mut();
-        Ok(world.get_entities())
-    }
-
-    fn is_entity_alive(&self, entity_id: u32) -> bool {
-        let world = self.inner.borrow_mut();
-        world.is_entity_alive(entity_id)
-    }
-
-    fn set_mode(&self, mode: String) {
-        let mut world = self.inner.borrow_mut();
-        world.current_mode = mode;
-    }
-
-    fn get_mode(&self) -> String {
-        let world = self.inner.borrow_mut();
-        world.current_mode.clone()
-    }
-
-    fn get_available_modes(&self) -> Vec<String> {
-        let world = self.inner.borrow_mut();
-        world
-            .registry
-            .lock()
-            .unwrap()
-            .all_modes()
-            .into_iter()
-            .collect()
-    }
-
-    fn list_components(&self) -> Vec<String> {
-        let world = self.inner.borrow_mut();
-        world.registry.lock().unwrap().all_component_names()
-    }
-
-    fn get_component_schema(&self, name: String) -> PyResult<PyObject> {
-        let world = self.inner.borrow_mut();
-        if let Some(schema) = world.registry.lock().unwrap().get_schema_by_name(&name) {
-            let json = serde_json::to_value(&schema.schema)
-                .map_err(|e| PyValueError::new_err(e.to_string()))?;
-            Python::with_gil(|py| Ok(to_pyobject(py, &json)?.into()))
-        } else {
-            Err(PyValueError::new_err("Component schema not found"))
-        }
-    }
-
-    fn move_entity(&self, entity_id: u32, dx: f32, dy: f32) {
-        let mut world = self.inner.borrow_mut();
-        world.move_entity(entity_id, dx, dy);
-    }
-
-    fn move_all(&self, dx: i32, dy: i32) {
-        let mut world = self.inner.borrow_mut();
-        world.register_system(MoveAll {
-            delta: MoveDelta::Square { dx, dy, dz: 0 },
-        });
-        world.run_system("MoveAll", None).unwrap();
-    }
-
-    /// Add a cell to the map (only works for SquareGridMap).
-    fn add_cell(&self, x: i32, y: i32, z: i32) {
-        let mut world = self.inner.borrow_mut();
-        if let Some(map) = &mut world.map {
-            // Try downcasting to SquareGridMap
-            if let Some(square) = map
-                .topology
-                .as_any_mut()
-                .downcast_mut::<engine_core::map::SquareGridMap>()
-            {
-                square.add_cell(x, y, z);
-            }
-        }
-    }
-
-    fn damage_entity(&self, entity_id: u32, amount: f32) {
-        let mut world = self.inner.borrow_mut();
-        world.damage_entity(entity_id, amount);
-    }
-
-    fn damage_all(&self, amount: f32) {
-        let mut world = self.inner.borrow_mut();
-        world.register_system(DamageAll { amount });
-        world.run_system("DamageAll", None).unwrap();
-    }
-
-    fn tick(&self) {
-        let mut world = self.inner.borrow_mut();
-        world.run_system("MoveAll", None).unwrap();
-        world.run_system("DamageAll", None).unwrap();
-        world.run_system("ProcessDeaths", None).unwrap();
-        world.run_system("ProcessDecay", None).unwrap();
-        world.turn += 1;
-    }
-
-    fn get_turn(&self) -> u32 {
-        let world = self.inner.borrow_mut();
-        world.turn
-    }
-
-    fn process_deaths(&self) {
-        let mut world = self.inner.borrow_mut();
-        world.register_system(ProcessDeaths);
-        world.run_system("ProcessDeaths", None).unwrap();
-    }
-
-    fn process_decay(&self) {
-        let mut world = self.inner.borrow_mut();
-        world.register_system(ProcessDecay);
-        world.run_system("ProcessDecay", None).unwrap();
-    }
-
-    fn count_entities_with_type(&self, type_str: String) -> usize {
-        let world = self.inner.borrow_mut();
-        world.count_entities_with_type(&type_str)
-    }
-
-    fn modify_stockpile_resource(&self, entity_id: u32, kind: String, delta: f64) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        world
-            .modify_stockpile_resource(entity_id, &kind, delta)
-            .map_err(PyValueError::new_err)
-    }
-
-    fn save_to_file(&self, path: String) -> PyResult<()> {
-        let world = self.inner.borrow_mut();
-        world
-            .save_to_file(std::path::Path::new(&path))
-            .map_err(|e| PyIOError::new_err(e.to_string()))
-    }
-
-    fn load_from_file(&mut self, path: String) -> PyResult<()> {
-        let registry = {
-            let world = self.inner.borrow_mut();
-            world.registry.clone()
-        };
-        let loaded = World::load_from_file(std::path::Path::new(&path), registry)
-            .map_err(|e| PyIOError::new_err(e.to_string()))?;
-        let mut world = self.inner.borrow_mut();
-        *world = loaded;
-        Ok(())
-    }
-
-    // --- System registration/bridge ---
-
-    fn register_system(&self, py: Python, name: String, callback: Py<PyAny>) -> PyResult<()> {
-        self.systems.register_system(py, name, callback)
-    }
-
-    fn run_system(&self, py: Python, name: String) -> PyResult<()> {
-        self.systems.run_system(py, name)
-    }
-
-    fn run_native_system(&self, name: String) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        world
-            .run_system(&name, None)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    // --- Event bus methods ---
-
-    fn send_event(&self, event_type: String, payload: String) -> PyResult<()> {
-        event_bus::send_event(event_type, payload)
-    }
-
-    fn poll_event(&self, py: Python, event_type: String) -> PyResult<Vec<PyObject>> {
-        event_bus::poll_event(py, event_type)
-    }
-
-    fn poll_ecs_event(&self, py: Python, event_type: String) -> PyResult<Vec<PyObject>> {
-        let mut world = self.inner.borrow_mut();
-        let events = world.take_events(&event_type);
-        Ok(events
-            .into_iter()
-            .map(|e| serde_pyobject::to_pyobject(py, &e).unwrap().into())
-            .collect())
-    }
-
-    fn update_event_buses(&self) {
-        event_bus::update_event_buses()
-    }
-
-    // --- User input ---
-
-    fn get_user_input(&self, py: Python, prompt: String) -> PyResult<String> {
-        let builtins = py.import("builtins")?;
-        let input_func = builtins.getattr("input")?;
-        let result = input_func.call1((prompt,))?;
-        result.extract::<String>()
-    }
-
-    // --- Job system ---
-
-    #[pyo3(signature = (entity_id, job_type, **kwargs))]
-    fn assign_job(
-        &self,
-        entity_id: u32,
-        job_type: String,
-        kwargs: Option<&Bound<'_, PyDict>>,
-    ) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut job_val = serde_json::json!({
-            "job_type": job_type,
-            "status": "pending",
-            "progress": 0.0
-        });
-        if let Some(kwargs) = kwargs {
-            let extra: serde_json::Value = pythonize::depythonize(kwargs)?;
-            if let Some(obj) = extra.as_object() {
-                for (k, v) in obj {
-                    job_val[k] = v.clone();
-                }
-            }
-        }
-        world
-            .set_component(entity_id, "Job", job_val)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    fn register_job_type(&self, _py: Python, name: String, callback: Py<PyAny>) {
-        let mut world = self.inner.borrow_mut();
-        world.job_types.register_native(
-            &name,
-            Box::new(move |old_job, progress| {
-                Python::with_gil(|py| {
-                    let job_obj = to_pyobject(py, old_job).unwrap();
-                    let result = callback.call1(py, (job_obj, progress)).unwrap();
-                    serde_pyobject::from_pyobject(result.bind(py).clone()).unwrap()
-                })
-            }),
-        );
-    }
-
-    // --- Inventory ---
-
-    fn get_inventory(&self, py: Python<'_>, entity_id: u32) -> PyResult<Option<PyObject>> {
-        self.get_component(py, entity_id, "Inventory".to_string())
-    }
-
-    fn set_inventory(&self, entity_id: u32, value: Bound<'_, PyAny>) -> PyResult<()> {
-        self.set_component(entity_id, "Inventory".to_string(), value)
-    }
-
-    fn add_item_to_inventory(&self, entity_id: u32, item_id: String) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut inv = if let Some(val) = world.get_component(entity_id, "Inventory") {
-            val.clone()
-        } else {
-            serde_json::json!({})
-        };
-        let slots_opt = inv.get_mut("slots").and_then(|v| v.as_array_mut());
-        let slots = if let Some(slots) = slots_opt {
-            slots
-        } else {
-            inv["slots"] = serde_json::json!([]);
-            inv.get_mut("slots").unwrap().as_array_mut().unwrap()
-        };
-        slots.push(serde_json::Value::String(item_id));
-        world
-            .set_component(entity_id, "Inventory", inv)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    fn remove_item_from_inventory(
-        &self,
-        _py: Python<'_>,
-        entity_id: u32,
-        index: usize,
-    ) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut inv = if let Some(val) = world.get_component(entity_id, "Inventory") {
-            val.clone()
-        } else {
-            return Err(PyValueError::new_err("No Inventory component found"));
-        };
-        if let Some(slots) = inv.get_mut("slots").and_then(|v| v.as_array_mut()) {
-            if index < slots.len() {
-                slots.remove(index);
-                world
-                    .set_component(entity_id, "Inventory", inv)
-                    .map_err(|e| PyValueError::new_err(e.to_string()))
-            } else {
-                Err(PyValueError::new_err("Index out of bounds"))
-            }
-        } else {
-            Err(PyValueError::new_err("No slots array in Inventory"))
-        }
-    }
-
-    fn get_equipment(&self, py: Python<'_>, entity_id: u32) -> PyResult<PyObject> {
-        let world = self.inner.borrow();
-        use pyo3::types::{PyDict, PyTuple};
-        if let Some(val) = world.get_component(entity_id, "Equipment") {
-            let any = serde_pyobject::to_pyobject(py, val)?;
-            if let Ok(dict) = any.downcast::<PyDict>() {
-                if let Ok(Some(slots_any)) = dict.get_item("slots") {
-                    if let Ok(slots) = slots_any.downcast::<PyDict>() {
-                        for (k, v) in slots.iter() {
-                            if v.is_instance_of::<PyTuple>() && v.len().unwrap_or(1) == 0 {
-                                slots.set_item(k, py.None())?;
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(any.into())
-        } else {
-            Ok(PyDict::new(py).into())
-        }
-    }
-
-    fn equip_item(&self, entity_id: u32, item_id: String, slot: String) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-
-        // 1. Check inventory
-        let inv = world
-            .get_component(entity_id, "Inventory")
-            .ok_or_else(|| PyValueError::new_err("Entity has no inventory"))?;
-        let slots = inv
-            .get("slots")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| PyValueError::new_err("No slots array in Inventory"))?;
-        if !slots
-            .iter()
-            .any(|v| v == &serde_json::Value::String(item_id.clone()))
-        {
-            return Err(PyValueError::new_err("Item not in inventory"));
-        }
-
-        // 2. Check item metadata
-        let mut found = None;
-        for item_eid in world.get_entities_with_component("Item") {
-            if let Some(item_comp) = world.get_component(item_eid, "Item") {
-                if item_comp.get("id") == Some(&serde_json::Value::String(item_id.clone())) {
-                    found = Some(item_comp);
-                    break;
-                }
-            }
-        }
-        let item_meta = found.ok_or_else(|| PyValueError::new_err("Item not found"))?;
-
-        // 3. Check slot compatibility
-        let valid_slot = item_meta
-            .get("slot")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| PyValueError::new_err("Item missing slot info"))?;
-        if valid_slot != slot {
-            return Err(PyValueError::new_err("invalid slot"));
-        }
-
-        // 4. Get or create Equipment component with correct structure
-        let mut equipment = if let Some(val) = world.get_component(entity_id, "Equipment") {
-            val.clone()
-        } else {
-            let mut map = serde_json::Map::new();
-            map.insert("slots".to_string(), serde_json::json!({}));
-            serde_json::Value::Object(map)
-        };
-
-        // 5. Ensure "slots" is an object
-        let slots_obj = equipment
-            .get_mut("slots")
-            .and_then(|v| v.as_object_mut())
-            .ok_or_else(|| PyValueError::new_err("Equipment slots must be an object"))?;
-
-        // 6. Check if slot is already occupied
-        if let Some(existing) = slots_obj.get(&slot) {
-            if !existing.is_null() {
-                return Err(PyValueError::new_err("already equipped"));
-            }
-        }
-
-        // 7. Equip
-        slots_obj.insert(slot.clone(), serde_json::Value::String(item_id.clone()));
-        world
-            .set_component(entity_id, "Equipment", equipment)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    fn unequip_item(&self, entity_id: u32, slot: String) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut equipment = world
-            .get_component(entity_id, "Equipment")
-            .ok_or_else(|| PyValueError::new_err("No Equipment component"))?
-            .clone();
-        let slots_obj = equipment
-            .get_mut("slots")
-            .and_then(|v| v.as_object_mut())
-            .ok_or_else(|| PyValueError::new_err("Equipment slots must be an object"))?;
-        slots_obj.insert(slot, serde_json::Value::Null);
-        world
-            .set_component(entity_id, "Equipment", equipment)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    /// Returns a list of entity IDs assigned to the given region ID.
-    fn get_entities_in_region(&self, region_id: String) -> Vec<u32> {
-        let world = self.inner.borrow();
-        world.entities_in_region(&region_id)
-    }
-
-    /// Returns a list of entity IDs assigned to regions of the given kind.
-    fn get_entities_in_region_kind(&self, kind: String) -> Vec<u32> {
-        let world = self.inner.borrow();
-        world.entities_in_region_kind(&kind)
-    }
-
-    /// Returns a list of cells (as Python objects) assigned to the given region ID.
-    fn get_cells_in_region(&self, py: Python, region_id: String) -> PyResult<PyObject> {
-        let world = self.inner.borrow();
-        let cells = world.cells_in_region(&region_id);
-        Ok(serde_pyobject::to_pyobject(py, &cells)?.into())
-    }
-
-    /// Returns a list of cells (as Python objects) assigned to regions of the given kind.
-    fn get_cells_in_region_kind(&self, py: Python, kind: String) -> PyResult<PyObject> {
-        let world = self.inner.borrow();
-        let cells = world.cells_in_region_kind(&kind);
-        Ok(serde_pyobject::to_pyobject(py, &cells)?.into())
-    }
-
-    fn get_body(&self, py: Python<'_>, entity_id: u32) -> PyResult<Option<PyObject>> {
-        self.get_component(py, entity_id, "Body".to_string())
-    }
-
-    fn set_body(&self, entity_id: u32, value: Bound<'_, PyAny>) -> PyResult<()> {
-        self.set_component(entity_id, "Body".to_string(), value)
-    }
-
-    fn add_body_part(&self, entity_id: u32, part: Bound<'_, PyAny>) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut body = if let Some(val) = world.get_component(entity_id, "Body") {
-            val.clone()
-        } else {
-            serde_json::json!({})
-        };
-        let parts = body.get_mut("parts").and_then(|v| v.as_array_mut());
-        let parts = if let Some(parts) = parts {
-            parts
-        } else {
-            body["parts"] = serde_json::json!([]);
-            body.get_mut("parts").unwrap().as_array_mut().unwrap()
-        };
-        let part_json: Value = from_pyobject(part)?;
-        parts.push(part_json);
-        world
-            .set_component(entity_id, "Body", body)
-            .map_err(|e| PyValueError::new_err(e.to_string()))
-    }
-
-    fn remove_body_part(&self, entity_id: u32, part_name: String) -> PyResult<()> {
-        let mut world = self.inner.borrow_mut();
-        let mut body = if let Some(val) = world.get_component(entity_id, "Body") {
-            val.clone()
-        } else {
-            return Err(PyValueError::new_err("No Body component found"));
-        };
-
-        fn remove_part_recursive(parts: &mut Vec<Value>, name: &str) -> bool {
-            let mut i = 0;
-            while i < parts.len() {
-                let part = &mut parts[i];
-                if part.get("name").and_then(|n| n.as_str()) == Some(name) {
-                    parts.remove(i);
-                    return true;
-                }
-                if let Some(children) = part.get_mut("children").and_then(|v| v.as_array_mut()) {
-                    if remove_part_recursive(children, name) {
-                        return true;
-                    }
-                }
-                i += 1;
-            }
-            false
-        }
-
-        if let Some(parts) = body.get_mut("parts").and_then(|v| v.as_array_mut()) {
-            if remove_part_recursive(parts, &part_name) {
-                world
-                    .set_component(entity_id, "Body", body)
-                    .map_err(|e| PyValueError::new_err(e.to_string()))
-            } else {
-                Err(PyValueError::new_err("Body part not found"))
-            }
-        } else {
-            Err(PyValueError::new_err("No parts array in Body"))
-        }
-    }
-
-    fn get_body_part(
-        &self,
-        py: Python<'_>,
-        entity_id: u32,
-        part_name: String,
-    ) -> PyResult<Option<PyObject>> {
-        let world = self.inner.borrow();
-        if let Some(body) = world.get_component(entity_id, "Body") {
-            fn find_part<'a>(parts: &'a [Value], name: &str) -> Option<&'a Value> {
-                for part in parts {
-                    if part.get("name").and_then(|n| n.as_str()) == Some(name) {
-                        return Some(part);
-                    }
-                    if let Some(children) = part.get("children").and_then(|v| v.as_array()) {
-                        if let Some(found) = find_part(children, name) {
-                            return Some(found);
-                        }
-                    }
-                }
-                None
-            }
-            if let Some(parts) = body.get("parts").and_then(|v| v.as_array()) {
-                if let Some(part) = find_part(parts, &part_name) {
-                    let py_obj = to_pyobject(py, part)?;
-                    return Ok(Some(py_obj.into()));
-                }
-            }
-        }
-        Ok(None)
-    }
-}
+pub use crate::python_api::PyWorld;
