+++ engine_py/src/python_api/mod.rs
@@ -1,0 +1,10 @@
+pub mod body;
+pub mod component;
+pub mod entity;
+pub mod equipment;
+pub mod inventory;
+pub mod misc;
+pub mod region;
+pub mod world;
+
+pub use world::PyWorld;
